Update README.md
Removed items from future work to represent completed functionality. Added a section relevant to functionality that was discussed as a possibility. Changed or added to parts of the opening sections to reflect changes to the operation of the system, specifically accessing the hosted site and retrieving signatures from the database.